// src/graphql/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };

export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  JSON: any;
  RegExpAsString: any;
};

export type AbilityScore = {
  __typename?: "AbilityScore";
  _id: Scalars["String"];
  desc?: Maybe<Array<Maybe<Scalars["String"]>>>;
  full_name?: Maybe<Scalars["String"]>;
  index?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
  skills?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
  url?: Maybe<Scalars["String"]>;
};

export type AbilityScoreSkills = {
  __typename?: "AbilityScoreSkills";
  index?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
  url?: Maybe<Scalars["String"]>;
};
// ...
export type Class = {
  __typename?: "Class";
  _id: Scalars["String"];
  class_levels?: Maybe<Scalars["String"]>;
  hit_die?: Maybe<Scalars["Float"]>;
  index?: Maybe<Scalars["String"]>;
  name?: Maybe<Scalars["String"]>;
  proficiencies?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
  proficiency_choices?: Maybe<Array<Maybe<ClassProficiency_Choices>>>;
  saving_throws?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
  spellcasting?: Maybe<ClassSpellcasting>;
  spells?: Maybe<Scalars["String"]>;
  starting_equipment?: Maybe<Array<Maybe<ClassStarting_Equipment>>>;
  starting_equipment_options?: Maybe<
    Array<Maybe<ClassStarting_Equipment_Options>>
  >;
  subclasses?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
  url?: Maybe<Scalars["String"]>;
};

export type ClassProficiency_Choices = {
  __typename?: "ClassProficiency_choices";
  choose?: Maybe<Scalars["Float"]>;
  from?: Maybe<Array<Maybe<AbilityScoreSkills>>>;
  type?: Maybe<Scalars["String"]>;
};

export type ClassSpellcasting = {
  __typename?: "ClassSpellcasting";
  info?: Maybe<Array<Maybe<ClassSpellcastingInfo>>>;
  level?: Maybe<Scalars["Float"]>;
  spellcasting_ability?: Maybe<AbilityScoreSkills>;
};

export type ClassSpellcastingInfo = {
  __typename?: "ClassSpellcastingInfo";
  desc?: Maybe<Array<Maybe<Scalars["String"]>>>;
  name?: Maybe<Scalars["String"]>;
};

export type ClassStarting_Equipment = {
  __typename?: "ClassStarting_equipment";
  equipment?: Maybe<AbilityScoreSkills>;
  quantity?: Maybe<Scalars["Float"]>;
};

export type ClassStarting_Equipment_Options = {
  __typename?: "ClassStarting_equipment_options";
  choose?: Maybe<Scalars["Float"]>;
  from?: Maybe<Array<Maybe<ClassStarting_Equipment_OptionsFrom>>>;
  type?: Maybe<Scalars["String"]>;
};

export type ClassStarting_Equipment_OptionsFrom = {
  __typename?: "ClassStarting_equipment_optionsFrom";
  equipment?: Maybe<AbilityScoreSkills>;
  quantity?: Maybe<Scalars["Float"]>;
};